Optimizations : adding controllers, separating middleware, adding module, lazy loading, using global config, node enrironment variables.
