+++ backend/app.js
@@ -9,7 +9,11 @@ const userRoutes = require('./routes/user');
 const app = express();
 
 mongoose
-  .connect('mongodb+srv://dilma:iC8YOwZQbkzvmWE1@cluster0.d4hhbv3.mongodb.net/node-angular?retryWrites=true&w=majority')
+  .connect(
+    "mongodb+srv://dilma:" + 
+    process.env.MONGO_ATLAS_PW + 
+    "@cluster0.d4hhbv3.mongodb.net/node-angular?retryWrites=true&w=majority"
+  )
   .then(() => {
     console.log('Connected to database!');
   })
