+++ backend/middleware/check-auth.js
@@ -3,7 +3,7 @@ const jwt = require('jsonwebtoken');
 module.exports = (req, res, next) => {
     try {
         const token = req.headers.authorization.split(" ")[1];
-        const decodedToken = jwt.verify(token, 'secret_this_should_be_longer');
+        const decodedToken = jwt.verify(token, process.env.JWT_Key);
         req.userData = { email: decodedToken.email, userId: decodedToken.userId };
         next();
     } catch (error) {
