+++ nodemon.json
@@ -1,0 +1,6 @@
+{
+    "env": {
+        "MONGO_ATLAS_PW": "iC8YOwZQbkzvmWE1",
+        "JWT_KEY": "secret_this_should_be_longer"
+    }
+}
