+++ src/app/auth/auth-routing.module.ts
@@ -1,0 +1,22 @@
+import { NgModule } from "@angular/core";
+import { RouterModule } from "@angular/router";
+
+import { LoginComponent } from "./login/login.component";
+import { SignupComponent } from "./signup/signup.component";
+
+const routes = [
+    { path: 'login', component: LoginComponent },
+    { path: 'signup', component: SignupComponent }
+]
+
+
+@NgModule({
+    imports: [
+        RouterModule.forChild(routes)
+    ],
+    exports: [
+        RouterModule
+    ]
+})
+
+export class AuthRoutingModule {}
