+++ src/app/auth/auth.guard.ts
@@ -16,7 +16,7 @@ export class AuthGuard implements CanActivate {
         const isAuth = this.authService.getIsAuth();
         
         if (!isAuth) {
-            this.router.navigate(['/login']);
+            this.router.navigate(['/auth/login']);
         }
         return isAuth;
     }
