+++ src/app/auth/auth.module.ts
@@ -1,0 +1,23 @@
+import { CommonModule } from "@angular/common";
+import { NgModule } from "@angular/core";
+import { ReactiveFormsModule } from "@angular/forms";
+
+import { AngularMaterialModule } from "../angular-material.module";
+import { AuthRoutingModule } from "./auth-routing.module";
+import { LoginComponent } from "./login/login.component";
+import { SignupComponent } from "./signup/signup.component";
+
+@NgModule({
+    declarations: [
+        LoginComponent,
+        SignupComponent,
+    ],
+    imports: [
+        CommonModule,
+        ReactiveFormsModule,
+        AngularMaterialModule,
+        AuthRoutingModule
+    ]
+})
+
+export class AuthModule {}
