+++ src/app/auth/auth.service.ts
@@ -3,7 +3,10 @@ import { Injectable } from "@angular/core";
 import { Router } from "@angular/router";
 import { Subject } from "rxjs";
 
+import { environment } from "../../environments/environment";
 import { AuthData } from "./auth-data.model";
+
+const BACKEND_URL = environment.apiUrl + "/user/";
 
 @Injectable({ 
     providedIn: 'root' 
@@ -38,7 +41,7 @@ export class AuthService {
 
     public createUser(email: string, password: string) {
         const authData: AuthData = {email: email, password: password};
-        this.http.post('http://localhost:3000/api/user/signup', authData)
+        this.http.post(BACKEND_URL + 'signup', authData)
         .subscribe({
             next: (response) => {
               this.router.navigate(['/']);
@@ -52,7 +55,7 @@ export class AuthService {
     public login(email: string, password: string) {
         const authData: AuthData = {email: email, password: password};
         this.http
-            .post<{ message: string, token: string, expiresIn: number, userId: string }>('http://localhost:3000/api/user/login', authData)
+            .post<{ message: string, token: string, expiresIn: number, userId: string }>(BACKEND_URL + 'login', authData)
             .subscribe({
                 next: (response) => {
                     const token = response.token;
