+++ src/app/posts/posts.module.ts
@@ -1,0 +1,28 @@
+import { NgModule } from "@angular/core";
+import { ReactiveFormsModule } from "@angular/forms";
+import { CommonModule } from "@angular/common";
+import { RouterModule } from "@angular/router";
+
+import { PostListComponent } from "./post-list/post-list.component";
+import { PostCreateComponent } from "./post-create/post-create.component";
+import { AngularMaterialModule } from "../angular-material.module";
+
+@NgModule({
+    declarations: [
+        PostListComponent,
+        PostCreateComponent
+    ],
+    imports: [
+        CommonModule,
+        ReactiveFormsModule,
+        AngularMaterialModule,
+        RouterModule
+    ],
+    exports: [
+        PostListComponent,
+        PostCreateComponent
+    ]
+
+})
+
+export class PostsModule {}
