+++ src/app/posts/posts.service.ts
@@ -2,7 +2,11 @@ import { HttpClient } from "@angular/common/http";
 import { Injectable } from "@angular/core";
 import { Router } from "@angular/router";
 import { map, Subject } from "rxjs";
+
+import { environment } from "../../environments/environment";
 import { Post } from "./post.model";
+
+const BACKEND_URL = environment.apiUrl + "/posts/";
 
 @Injectable({ 
     providedIn: 'root' 
@@ -18,7 +22,7 @@ export class PostsService {
 
     getPosts(postsPerpage: number, currentPage: number) {
         const queryParams = `?pagesize=${postsPerpage}&page=${currentPage}`;
-        this.http.get<{message: string, posts: any, maxPosts: number}>('http://localhost:3000/api/posts' + queryParams)
+        this.http.get<{message: string, posts: any, maxPosts: number}>(BACKEND_URL + queryParams)
         .pipe(
             map((postData) => {
                 return { posts: postData.posts.map(post => {
@@ -45,7 +49,7 @@ export class PostsService {
             content: string, 
             imagePath: string,
             creator: string
-        }>('http://localhost:3000/api/posts/' + id);
+        }>(BACKEND_URL + id);
     }
 
     getPostUpdateListener() {
@@ -58,7 +62,7 @@ export class PostsService {
         postData.append('content', content);
         postData.append('image', image, title);
 
-        this.http.post<{ message: string, post: Post }>('http://localhost:3000/api/posts', postData)
+        this.http.post<{ message: string, post: Post }>(BACKEND_URL, postData)
         .subscribe((responseData) => {
             this.router.navigate(['/']);
         });
@@ -82,13 +86,13 @@ export class PostsService {
             };
         };
 
-        this.http.put('http://localhost:3000/api/posts/' + id, postData)
+        this.http.put(BACKEND_URL + id, postData)
         .subscribe((response) => {
             this.router.navigate(['/']);
         });
     }
 
     deletePost(postId: string) {
-       return this.http.delete('http://localhost:3000/api/posts/' + postId);
+       return this.http.delete(BACKEND_URL + postId);
     }
 }
