+++ src/environments/environment.ts
@@ -1,0 +1,5 @@
+
+export const environment = {
+    production: false,
+    apiUrl: "http://localhost:3000/api"
+};
